users_service/requirements.txt
flask
flask-jwt-extended
flask_sqlalchemy
psycopg2-binary
